+++ src/SMAPI/Framework/ContentPack.cs
@@ -59,6 +59,16 @@ namespace StardewModdingAPI.Framework
                 : null;
         }
 
+        /// <summary>Save data to a JSON file in the content pack's folder.</summary>
+        /// <typeparam name="TModel">The model type. This should be a plain class that has public properties for the data you want. The properties can be complex types.</typeparam>
+        /// <param name="path">The file path relative to the mod folder.</param>
+        /// <param name="data">The arbitrary data to save.</param>
+        public void WriteJsonFile<TModel>(string path, TModel data) where TModel : class
+        {
+            path = Path.Combine(this.DirectoryPath, PathUtilities.NormalisePathSeparators(path));
+            this.JsonHelper.WriteJsonFile(path, data);
+        }
+
         /// <summary>Load content from the content pack folder (if not already cached), and return it. When loading a <c>.png</c> file, this must be called outside the game's draw loop.</summary>
         /// <typeparam name="T">The expected data type. The main supported types are <see cref="Map"/>, <see cref="Texture2D"/>, and dictionaries; other types may be supported by the game's content pipeline.</typeparam>
         /// <param name="key">The local path to a content file relative to the content pack folder.</param>