+++ src/SMAPI/IContentPack.cs
@@ -22,10 +22,16 @@ namespace StardewModdingAPI
         ** Public methods
         *********/
         /// <summary>Read a JSON file from the content pack folder.</summary>
-        /// <typeparam name="TModel">The model type.</typeparam>
+        /// <typeparam name="TModel">The model type. This should be a plain class that has public properties for the data you want. The properties can be complex types.</typeparam>
         /// <param name="path">The file path relative to the content pack directory.</param>
         /// <returns>Returns the deserialised model, or <c>null</c> if the file doesn't exist or is empty.</returns>
         TModel ReadJsonFile<TModel>(string path) where TModel : class;
+
+        /// <summary>Save data to a JSON file in the content pack's folder.</summary>
+        /// <typeparam name="TModel">The model type. This should be a plain class that has public properties for the data you want. The properties can be complex types.</typeparam>
+        /// <param name="path">The file path relative to the mod folder.</param>
+        /// <param name="data">The arbitrary data to save.</param>
+        void WriteJsonFile<TModel>(string path, TModel data) where TModel : class;
 
         /// <summary>Load content from the content pack folder (if not already cached), and return it. When loading a <c>.png</c> file, this must be called outside the game's draw loop.</summary>
         /// <typeparam name="T">The expected data type. The main supported types are <see cref="Map"/>, <see cref="Texture2D"/>, and dictionaries; other types may be supported by the game's content pipeline.</typeparam>